requirements.txt
fastapi
uvicorn[standard]
supabase
openai-whisper
pydantic
python-multipart
transformers
torch
python-docx
reportlab
python-dotenv